Classify DEL (U+007F) as a control character.
The [Unicode Character Database] considers DEL (U+007F) to be a control
character, as do [Rust's `char::is_control`],
[Rust's `char::is_ascii_control`], [Rust's `u8::is_ascii_control`],
[C's `iscntrl`] (in ASCII environments), and [POSIX's POSIX locale],
so report it with `execute` rather than `print`.

This deviates from [Paul Williams' ANSI parser state machine], which
[treats 7F as printable].

This is because on DEC terminals which support the feature, one can
use SS2 or LS2R to select an alternate character set in which 7F maps
to a printable character. However, in modern environements supporting
UTF-8, character-set shifts are unneeded and not widely supported.

For users that don't support character-set shifts, 7F is always a
control character, and reporting it via `execute` helps users avoid
accidentally treating it as a printable character.

[Unicode Character Database]: https://www.unicode.org/reports/tr44/
[Rust's `u8::is_ascii_control`]: https://doc.rust-lang.org/std/primitive.u8.html#method.is_ascii_control
[Rust's `char::is_ascii_control`]: https://doc.rust-lang.org/std/primitive.char.html#method.is_ascii_control
[Rust's `char::is_control`]: https://doc.rust-lang.org/std/primitive.char.html#method.is_control
[C's `iscntrl']: http://www.cplusplus.com/reference/cctype/iscntrl/
[POSIX's POSIX locale]: https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap07.html#tag_07_03_01_01

[Paul Williams' ANSI parser state machine]: https://vt100.net/emu/dec_ansi_parser
[treats 7F as printable]: https://vt100.net/emu/dec_ansi_parser#STGRO